sys-devel/bison liby.a should be installable through use flag (see bug#204416) Signed-off-by: Robert A <r_a@lavabit.com>

+++ sys-devel/bison/bison-2.3.ebuild
@@ -1,6 +1,6 @@
-# Copyright 1999-2007 Gentoo Foundation
+# Copyright 1999-2009 Gentoo Foundation
 # Distributed under the terms of the GNU General Public License v2
-# $Header: /var/cvsroot/gentoo-x86/sys-devel/bison/bison-2.3.ebuild,v 1.10 2007/12/27 19:29:06 vapier Exp $
+# $Header: 
 
 inherit toolchain-funcs flag-o-matic
 
@@ -11,7 +11,7 @@ SRC_URI="mirror://gnu/bison/${P}.tar.bz2"
 LICENSE="GPL-2"
 SLOT="0"
 KEYWORDS="alpha amd64 arm hppa ia64 m68k mips ppc ppc64 s390 sh sparc ~sparc-fbsd x86 ~x86-fbsd"
-IUSE="nls static"
+IUSE="nls static liby"
 
 DEPEND="nls? ( sys-devel/gettext )"
 
@@ -29,8 +29,9 @@ src_install() {
 	# This one is installed by dev-util/yacc
 	mv "${D}"/usr/bin/yacc{,.bison} || die
 
-	# We do not need this.
-	rm -r "${D}"/usr/lib* || die
+	if ! use liby; then
+		rm -r "${D}"/usr/lib* || die
+	fi
 
 	dodoc AUTHORS NEWS ChangeLog README OChangeLog THANKS TODO
 }
